Develop

+++ src/main/java/aagapp_backend/dto/LeaderboardDto.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -15,6 +17,6 @@ public class LeaderboardDto {
     private String pictureUrl;
     private Integer score;
     private Boolean isWinner;
-    private Double winningAmount;
+    private BigDecimal winningAmount;
 }
